+++ R/sigTests-methods.R
@@ -3,9 +3,3 @@
 setMethod("sigTests", "LM", function(object, ...) {
   object@sigTests
 })
-
-#' @describeIn sigTests Extracts a matrix with the results of significance
-#'   tests.
-setMethod("sigTests", "SummaryLM", function(object, ...) {
-  object@sigTests
-})